Switch from vector of hash-set to hash-map of vector in Partitioner::mask()
This changeset switches the data structure used between query() and
mask() in Partitioner. Previously query() prepared a vector of unordered
sets to store the potentially multiple (but unique) chare IDs associated
to mesh node IDs and sent that vector to mask() on another PE. The size
of the vector was always the same as the number mesh nodes arriving in
the id argument of query(), the same as was expected by mask(), but
since not all mesh chares contribute to a given mesh node queried by the
sender (only some of those that are on a PE boundary), the vector
previously potentially contained a significant number of empty sets.

This changeset changes the vector of unordered sets to an unordered map
of vectors where the map key stores the mesh node ID the associated
chare IDs in the vector contribute to. This way, we account for the fact
whether a list of chares contribute to a mesh node not by an empty
(unordered) set but by the key in the new unordered map and thus do not
have to store empty sets (or vectors). While this approach sends the
mesh node ID values back together with the chare IDs in the vectors
(which would not strictly be necessary, since the requestor knows about
those mesh node IDs already), this strategy compresses the data that
needs to be communicated since it does not send a potentially large
number of empty sets.

Rationale 1: The assumption behind this change is that there is a lot
less mesh node IDs to which chares contribute to on neighboring PEs than
the number of domain mesh nodes on a PE. This is usually the case, since
a domain decomposition (presumably) always yields more domain-nodes
compared to PE-domain-boundary nodes.

Rationale 2: Another important consideration is that the data-exchange
at this precise point during setup seems to be the bottle-neck. This is
most likely due to the potentially large amount of data that needs to be
communicated among all PEs. (1) This bottle-neck part of the algorithm
starts out, in gather() with a broadcast, query(), started on all PEs,
i.e., an all-to-all, sending all the domain-mesh-node IDs to all other
PEs for identifying the boundary nodes other PEs will contribute to
after mesh partitioning. An all-to-all with all domain-mesh-node IDs is
potentially a huge amount data movement to start out with. (2) After
searching through all mesh node IDs a PE holds, from query() each PE
sends the result of its search back to all other PEs. This step is only
a point-to-point message, but from all PEs to all other PEs. This cannot
be a broadcast, since the answer is different for each PE. This step
still may involve quite a large amount of communication, although most
likely less than the first step. As a consequence, any reduction in the
size of the messages (or the complexity of the data structures
communicated) of any of these two steps should yield less memory use and
better performance.

Note that since the type of Partitioner::m_id does not change, it
remains an (ordered) std::set, this changeset does not change the
algorithmic complexity of the search in query(), in step (2) discussed
above, it remains a log(n)-search for all domain mesh node IDs the
requestor PE is querying. Another potential improvement on this
algorithm therefore is to replace m_id with an unordered set, which
facilitates O(1) searches.

What does change, however, is the way the resulting data structure, now
an unordered map storing vectors, is used. In mask() we now do not have
to loop through m_id (storing all domain mesh node IDs) and only deal
with boundary mesh node IDs that the sender PE has matched, but only
loop through the matched node IDs. This happens twice in mask(): (1) in
computing m_msum, and (2) in computing the temporary communication maps
in m_comm.

The inner data structure (unordered set) is replaced by a vector, since
there is no need for searching in that data in any way, only uniqueness
is required, which is now guaranteed by calling tk::unique() on each
chare ID list (now stored as a vector) after it was constructed in
flatten().

As measured on 720 cores with a 52M-cell mesh, this change amounts to a
modest performance increase: 15s instead of 17s of total runtime (setup
+ 1 time step).