Put in capability to estimate L1-norm in inciter
For those PDE problems where an analytical solution is available, this
commit puts in a way to estimate the L1 (absolute value) norm that is
proportional to the error for all components computed.

We piggy-back on the diagnostics calculation and communication as well
as on the PDE::initialize() routine, since that assigns the analytical
solution for those PDE problems that have analytical solution defined
and available.