First step towards initial uniform refinement
This is work in progress towards a capability to do uniform (1:8)
refinement right after mesh reordering. This is an off-shoot attempt and
not necessarily intended to be final, even if works, since the initial
uniform refinement should probably be a specific case of fully automatic
refinement, already underway and coming later, enabling code reuse.

Beside being useful by itself for initial uniform refinement, the main
purpose of this off-shoot attempt is to quickly be able to do scaling
studies on very large meshes without having to read a large mesh
initially. As @junghans points this out, this can also be very handy for
debugging later as large meshes do not have to be sent for testing.

What works so far is the uniform 1:8 refinement, i.e., all tetrahedra in
the mesh that are read in initially are replaced with 8 new cells: (1) a
new mesh node correctly is introduced on all edges to its halfway
location, and (2) 8 new mesh cells, i.e., their correct connectivities,
are introduced in place of every original mesh cell. This works in
serial for now.

Some data structures, derived from the initial mesh connectivity (read
from file) and some of the algorithms that use them, are already made
aware of the refined mesh but some are not yet. For example, the correct
refined mesh is written out to file, but the new->old mesh node ID map
is not yet aware of the new mesh nodes and thus the boundary conditions
lookup does not work yet. Similarly finding PEs to associated to the
newly introduced mesh nodes does not work yet. Another thing missing is
the communication of the newly introduces mesh nodes that the different
PEs must be made aware of.