WIP towards reconnecting data structures after initial uniform refinement
Details:

 * Refinement is commented out, inciter regression tests pass.

 * When refinement is enabled, we now generate a new data structure that
associates edges (a pair of old node IDs) to new node IDs categorized by
chares for only the nodes newly added as a result of initial uniform
refinement. This will be used for looking up boundary conditions from
file and matching to newly added node IDs.

 * The four data structures required by firing up workers, generated in
Partitioner::reordered(), are m_chcid, m_node, m_msum, and m_id, and the
now newly added m_chceid, only nonempty if refinement has been
performed. Note that m_id is only needed for the diagnostic cost
calculation and not passed to workers (Carrier).

 * At this point, refinement, though partially works, can be simply
commented out, and it has no effect on the four data structures derived
from mesh connectivity. This is verified by making sure the inciter
regression tests pass.

Next is to continue with *using* these data structures in Carrier and
fixing them and the client code up so that the refinement mesh is
computed on.