Initial uniform refinement appears to work in serial
Now we not only have a correct refined mesh but the solver computes
reasonable-looking results on it - without applying boundary conditions.

Details:

* Refinement, which will be made optional, is now not commented out.

* The full mesh node coordinates are still passed to the workers, which
is only correct in serial.

* Partitioner::m_msum now holds sets instead of vectors. This is so that
we can efficiently search for old node IDs in them and augment them with
the newly added nodes if an edge (with two endpoints as old node IDs) is
found. msum, mesh surrounding mesh (chunks), is used for communicating
among workers throughout the algorithms in Carrier (volumes, FCT, etc.).
Note that the sets in msum are converted to vectors before passing them
to Carrier's constructors as they are easier to use them as vectors
there.

* Also updated the child element connectivities of the 8 new mesh cells
replacing the cells read from file, which now have positive cell
volumes. Added assert to places where the Jacobian is computed to
trigger error on negative cell volumes. (We should also add basic mesh
quality checks in the future, but these asserts are okay to stay.)

* The communication cost calculation has also been updated with the
newly added nodes.

Next steps:

* Modify Carrier to assign BCs to newly added nodes.

* Make it work in parallel too, which will require interrogating
Partitioner::m_communication and send/receive the node IDs of the newly
added nodes.

* Replace the mesh coordinate read to now only be in Partitioner and
take it out from Carrier.