Make sure tests fail if exo files are different
In case of multiple exodiff tests inside a single regression test
cmake's ctest has so far only been looking for a fail regular expression
for 'exodiff: ERROR', but not for 'exodiff: Files are different'. Since
the former is triggered by an exodiff error only, regression tests that
do not fail exodiff with an error but the output of the test was 'file
different', did not trigger a failed test for ctest. This commit fixes
this problem.