Assign BCs to newly added nodes during initial uniform refinement
Assigning Dirichlet BCs to newly added mesh nodes during initial uniform
refinement appears to work. Still serial only.

Details:

The new BC-node coordinates have been tested refining two different
meshes using both manual inspection as well as outputing the coordinates
for a side set known to be aligned with a coordinate plane.

Adding new nodes is now done without relying on the generation of unique
edges surrounding all mesh nodes using tk::genEdsup(). For some reason,
the algorithm in tk::genEdsup() does not seem to generate all unique
edges. This may be a bug in tk::genEdsup(), but may also be my erroneous
expectation of that algorithm. In particular, the map 'star' in
tk::genEdsup() appears correct, however, its conversion to the linked
list, {edsup1,edsup2}, data structure or its access does not appear to
produce the same results as in the map. This would need some further
investigation to fix. Though tk::genEdsup() is unit-tested, only with a
simple mesh that was manually verified at the time the test was added.
The algorithm may not be correct with other meshes. In general, one
could also argue, that the algorithms, including genEdsup, in
Mesh/DerivedData.C are of somewhat limited value for dynamically chaning
data as they return linked lists. These lists are known to only perform
well for static data, i.e., frequently and dynamically changing them
will not be optimal. For dynamically changing data, for such data
structures derived from the mesh connectivity, hash-maps, ordered maps,
and sets will be better as they support O(log(n)) and O(1) complexity
searches. Perhaps, these algorithms should be also written producing
such dynamic data structures as well in the near future, which appears
to be required for adaptive mesh refinement anyway. Bottom line: for the
current work of getting uniform initial refinement working, we no longer
rely on genEdsup(), but pulled our own in Partitioner::refine(), which
is not only simpler, but yields a hash-map for the unique edges and thus
easier to use, not to mention that it avoids the data conversion from
map to linked list so it probably uses less memory too.  Nevertheless,
the derived data structure algorithms should (and most likely soon will)
be revisited for dynamic data.

Next is parallelization of the uniform initial mesh refinement.