WIP towards parallel uniform refinement
Details:

* At this point the refinement is commented out and the behavior appears
correct in parallel.

* Switched key and value in edgenodemap passed to Carrier as it is only
the inverse that was needed there and in Partitioner it does not really
matter which one we generate as we don't have to search in it (only in
Carrier).

* The Carrier chares still read the mesh node coordinates. This should
probably be replaced by sending the correct coordinates from
Partitioner.

* Add a bunch of code towards the update of the IDs of newly added nodes
during uniform refinement across PEs. Though this builds, it does not
work yet and commented out for now and is being debugged.

* Partitioner::m_msum for now is reverted to hold vectors instead of
sets. It would be good if we could get away with vector.