WIP towards parallel uniform refinement
Compared to e49d45c2e23 we now have uniform initial mesh refinement that
works in serial. In other words, the initial uniform refinement is now
NOT commented out and it works in serial. Previously, e49d45c2e23 worked
in parallel but without refinement.

Details, background, design:

* At e49d45c2e23 I did not yet have a good idea how the newly-generated
edge-nodes will be transferred from Partitioner to Carrier. Previous
work on initial uniform refinement simply punted on the mesh node
coordinates, and though new edge-nodes were added, their coordinates
were simply passed to Carriers or were not passed at all and only the
old (un-refined) mesh nodes were re-read there. Those are both punts,
since the former is only correct in serial, while the latter is only
correct without refinement.

* At first, I wanted to pass both the old and the newly generated node
coordinates (in a single coordinates data structure) to Carrier, instead
of having to read the old ones ("again") there, but that has turned out
to be a little problematic due to mesh node reordering. That is because
due to the reordering the original mesh-node coordinates that are read
in parallel in chunks by PEs are not the same as needed for the Carrier
chares spawned by the PE, precisely because they are now globally
reordered, i.e., new global node IDs have been assigned to them. This
reordering happens regardless of whether initial uniform refinement has
been performed or not. As a result, the reordered node IDs do not all
exist in the chunk of the mesh read on a PE. Therefore, making the node
coordinates available to the Carrier chares are not possible without an
additional communication step. While an additional communication step
may be preferable and more performant, for now (as before) we throw away
the mesh node coordinates originally read by Partitioner and in Carrier
we re-read them from the mesh file.

* The newly added edge-nodes, during initial uniform refinement, are
recreated in the same way as in Partitioner. Also, just noticing that,
since Partitioner's newly added edge-node coordinates are thrown away,
we don't even need to compute them. As this looks right now, this is
only needed in Carrier. (If these new edge-nodes will be communicated
instead, it will only be needed in Partitioner. In either case, this
should only be done either in Partitioner or in Carrier, but not both as
it is currently done.)

* The details of the Tet-map are now moved to UnsMesh to de-clutter
somewhat Partitioner::refine().

Next steps:

1. Debug the parallel edge-node ID communication before creating the
workers.

2. Update msum, required for parallel communication during time
stepping.

3. We could also explore switching from re-reading node coordinates from
file to communicating them instead. However, it is hard to judge the
difference in performance between these approaches at this point. Note
that re-reading the same mesh nodes (by different PEs) may not be as I/O
intensive as it sounds, as some of that data might be in cache already
as it has been read before. Also, going with communication instead of
re-read might entail communication in not only one but two steps: (1)
definitely among Partitioner PEs, and (2) perhaps also when doling out
work to Carrier worker chares, which may or may not necessarily always
be created on the same PE by the runtime system and this latter can also
be influenced by the degree of virtualization. In view of these, which
algorithm will perform better is hard to estimate  at this point. It
should be measured and compared and since this requires more code and
the current solution already works (in serial) and appears correct, we
leave that "excercise" for another day.