Re-enable msum
This commit re-enables the calculation of msum which stores the
per-chare maps for communicating node IDs during time stepping. With
this we can now run in parallel though still without turning on initial
uniform mesh refinement.

Here is where we are:

                without initial   with initial
                  refinement       refinement
--------------------------------------------------
serial             works              works
parallel           works             not yet

Next is to debug the asynchronous update of the IDs of newly added
edge-nodes during initial refinement.