Add back gnu macros to Intel build
Previously we used -gcc-sys with the Intel compiler, which only defines
GNU macros during the compilation of system headers. However, since that
is not the default, it causes problems with certain versions of the GNU
libstdc++ headers.

The real reason of using -gcc-sys in the first place was to not trigger
a bunch of compiler warnings originating from 'pragma diagnostic
push/pop' lines in src/NoWarning/*, which selectively ignore various
compiler warnings from third-party libraries, because the Intel compiler
also defines the __GNUC__ macro and that is also how gcc was detected,
but the Intel compiler, even though it should recognize the GCC
diagnostic push/pop pragmas, it does not, so it ends up generating
warnings.

This commit takes the -gcc-sys argument out of the Intel compiler
arguments, thus reverts this to the default, and so the GNU macros are
now defined not only for system files (which seem required, otherwise
they don't always compile with the Intel compilers), but also for our
source files. So this would trigger the warnings, but  this is now
ignored by a smarter compiler macro that can strictly detect the gnu
compiler and not the Intel or clang both of which define gnu macros.
This new STRICT_GNUC macro is defined in Base/Macro.h.

This commit is now also tested using the Intel MPI using the Intel
compilers. Previously the Intel compilers were only used in conjunction
with OpenMPI on linux.

Using the Intel MPI with the Intel compilers is done by loading the
Intel-provided 'mpi' module file, setting paths, etc., + setting the
following shell environment variables:

I_MPICXX=icpc
I_MPICC=icc
I_MPIFC=ifort
I_MPIF77=ifort
I_MPIF90=ifort

These variables influence IntelMPI's mpicxx, mpicc, etc. wrappers to use
the Intel compilers instead of their default gnu equivalents. After
this, our TPL and code cmake runs pick up the correct wrappers as well
as the underlying compilers.