Understand errors and convergence rate.
This commit collects a bunch of seemingly unrelated changes as a result
of my increased understanding of how the computed numerical errors in
inciter correspond to the convergence rate for smooth problems using
initial uniform mesh refinement to generate fine meshes from coarse
meshes (and using those in a convergence study).

In a nutshell, doing a simple convergence study, e.g., solving the pure
advection equation on the slotted cylinder problem (only advecting the
smooth Gaussian hump), one can compute the convergence rate as p = log2(
Ec / Ef ), where Ec and Eh are the errors computed for the coarse and
the fine resolution meshes, respectively. In this expression, it is
assumed that the largest edge length of the fine mesh is exactly twice
as that of the coarse mesh and that the numerical error is proportional
to C * h^p, where C is some constant, h is the edge length, and p is the
order of convergence. If one solves with and without initial uniform
refinement as the fine and coarse meshes, respectively, however, the
largest edge-length of the fine mesh is NOT twice as long as that of the
coarse mesh (as I have originally assumed), but h_f = sqrt(2)/2 * h_c,
where h_f is the fine-mesh largest edge-length and h_c is the
coarse-mesh largest edge-length. As a result, the convergence rate
becomes p = log2( Ec / Ef ) = log2( C h_c^p / C h_f^p ) = log2[ (2 /
sqrt(2))^p ] = p / 2. This means that we get half the convergence rate
compared to the case where h_f = 2h_c. Therefore, for 2-stage
Runge-Kutta, a measured p = 1, means p is really 2. Thanks to @nncarlson
for pointing this out.

Thus I am now confident that the implemented temporal and spatial
discretization gives p = 2 for smooth problems, as expected, independent
of whether the L1 or the L2 norm is used to compute the errors. For now
this is tested with the adv-diff problem only.

Details on the actual changes:

 * Add new operators to tk::Data:

   - New overloads to compound operator*= and operator*, taking a
     tk::real as the right hand side, besides the existing tk::Data, as
     member functions. This allows new operations involving tk::Data
     objects as: D *= r and D = A * r, where D,A are tk::Data and r is
     tk::real.

   - New operator*, taking a tk::real as the left hand side in D = r *
     A, where D,A are tk::Data and r is tk::real, as non-member function.

   - Add unit tests for the above operators.

 * Replace and simplify the computation of the error. Previously, the L1
   norm was computed across the domain, now we compute the L2 norm instead.
   Both work fine, and there is not much difference in FLOPS so we keep
   L2.

 * Take out computing the total volume. The precise mathematical
   definition of the errors (both L1 and L2) do not require dividing the
   final aggregated sum by the total volume (or its square-root for L2),
   and so this is taken out. This also requires less communication in
   parallel.

 * The multiplier that is different for different stages of a single
   time step, e.g., 1/2 and 1.0 for our 2-stage Runge-Kutta, is now not
   passed down in the PDE hierarchy, because this can simply be applied in
   Carrier::apply(), when the solution updates are applied. This also
   simplifies the interface of PDE::rhs() and thus affects all PDEs
   (Transport and Compflow for now).

 * The diagnostics, i.e., residuals and errors integrated over the whole
  computational domain now have a more detailed explanation in
  Carrier::diagnostics() that explains why and how this data is computed.

 * Remove debug-code from Compflow.h. Still have not decided how density
   will be updated for the vortical flow problem in CompFlowProblem.h:
   either update or not (in that problem the density does not change, so
   it should not matter, but it may affect the convergence rate). This
   will be tested later, so keep the current update as comments for now.

 * Class CompFlowProblemVorticalFlow now also outputs the pressure as
   a field variable.

 * Class Transport now outputs the L2 error of all components
   integrated.