Correct shear-diffusion test problem.
This commit puts in the necessary changes for a correct shear-diffusion
test problem. This test problem has an analytic solution (see
documentation for inciter::TransportProblemShearDiff) and so can be used
to test the numerical solution of the advection-diffusion equation in
3D.

Details:

 * Inciter now writes field output at the last time step, regardless of
whether it would have to write one (based on user settings) or not.
However, this write at the last time step only happens if there is new
data to be written out. See also Carrier::m_lastFieldWriteTime.

 * We now also divide the error computed across the whole domain by the
total mesh volume. This allows comparing the numerical value of the
error across different problems.

 * Carrier now also outputs the PDFs of the edge-lengths and the cubic
root of the cell volume. These are nice to look at when one does, e.g.,
a convergence study, in which the convergence is affected by the
definition of the edge length. The bin size for these PDFs are
hard-coded for now. We should come up for some better, mesh-independent
bin size in the future, but this is okay for now.

 * Also fix logic used to determine when to output field data.
Previously this was !(m_it % g_inputdeck.get< tag::interval, tag::field >())
and now it is !((m_it+1) % g_inputdeck.get< tag::interval,
tag::field >()). The second one is correct (as originally intended),
because it now outputs the field data of the currently finished time
step instead of the previous one. (Note that this should affect existing
regression tests, but does not since the field output interval has
always been 1 which is not affected by this change. The single
regression test for which the interval is larger than 1
(asynclogic/slot_cyl.q) does not test field output.

 * Fixed diffusion operator in TransportPhysicsAdvDiff, which now
correctly computes the diffusion coefficients for linear tetrahedra (1/6
was missing).

 * The diffusion operator can now also compute different diffusion
coefficients in different coordinate directions.

 * The time step size now also is correctly computed based on the
diffusion term in TransportPhysicsAdvDiff::diffusion_dt(). That is, dt
is now correctly proportional to dx^2/(2D) instead of dx^2/(D).
Additionally, the potentially different diffusion coefficients in
different directions are taken into account.

 * Also add documentation for TransportProblemShearDiff, showing the
equation, its analytic solution, and a link to the paper where the
solution is taken from.

 * Also, the solution is now the 3D solution taken from paper and not
the 2D-simplified version as it was previously.