Add L_{infinity} norm to inciter
Now besides the L2 norm, the Linf norm is also computed. This should
probably be configurable later (adding probably the L1 norm as well),
since different norms can be more advantageous depending on the problem.

Also discovered that the aggregation of the different norms in parallel
was not entirely correct. The correct way to compute the norms is to
visit all mesh nodes only once, but the chare-boundary nodes were
potential visited multiple times previous to this commit. Now in
LinSysMerger only the nodes owned are taken into consideration which is
exclusive so thus each node on the mesh are counted only once.

Since the Linf norm requires a different type of operation (max instead
of L2's sum) across PEs, this is now done using custom Charm++ reducers,
added in Inciter/DiagReducer.[Ch]. These declare and define
serialization and aggregation of a vector of vectors of real numbers
using different policies for the different norms.