Add artificial viscosity as a background damping for the Euler eqs
Details:

  * The user can set the value of the artificial viscosity in a
'compflow ...  end' block using the keyword 'artvisc', e.g., artvisc
1.5. If not given, the default is 1.0. Only used for the Euler equations
and only recognized in the compflow ... end block.

  * We now also check the for the positivity of the pressure when
computing the sound speed, required for computing time step size. If the
pressure is negative for whatever reason, we ignore it.

  * We now divide by 2*mu instead of just mu when computing the viscous
dt for Navier-Stokes - this is to be consistent with the time step
limits for the linear diffusion equation with explicit time stepping.

  * There is also a newly added Taylor-Green manufactured solution test
case, which is work in progress.