Add time-dependent Dirichlet BCs
Details:

* This change contains a considerable revamp of how Dirichlet boundary
conditions are queried from Carrier. The reason is that previously it
was only possible to assign a single constant value to a side set on
which Dirichlet boundary conditions can be set and the single value was
set for all nodes of a side set. Now it is possible to assign a
different BC value to every node on a given side set and those values
can also change in time. This is required by a verification test problem
(nonlinear energy growth and Rayleigh-Taylor) which have time-dependent
analytical solutions. These analytical solutions must be prescribed at
boundaries and these solutions along the boundary surfaces also change
with the spatial location. For user-defined problems the capability of
time-dependent Dirichlet BCs are also useful, though probably will be
only used in a less complex fashion with temporal but without spatial
dependence, i.e., a scalar-valued function of time assigned to all nodes
on a side set - though this functionality is only a plan at this point
as there is code to input and use this in a user-defined problem.

* As a result of this the way Carrier queries boundary conditions from
individual system of PDEs integrated had to change. In particular the
set of node IDs now also has to be passed into the PDE classes because
they are in the position to custom-evaluate and set their BCs for a
given side set and for a given problem configuration. This is used at
this point by the non-linear energy growth (NLEG) test problem.

* While the capability can handle more complex BCs, the code in Carrier
simplified, see Carrier::bc() and additional comments inline. A new map
is now pre-computed in Carrier's constructor as the BCs now are queried
at every time step stage to avoid re-computing some static data, see
Carrier::m_side.

* Another simplification is that the BCs are now not required for the
PDE::initialize() and all "children". See the absence of bc() in
Carrier::setup(). This also allows some simplification in the
LinSysMerger's asynchronous logic, see LinSys/linsysmerger.ci.

* Note the specific comments inline in Carrier::bc(), detailing of what
and how must be done when BC data structures are concatenated from
different PDE systems and then sent to LinSysMerger. Also note that the
algorithm in there will only work for a single PDE system at this point.

* Removed BCValMerger from LinSysMerger - was unused.

TODO:

* This commit is a step towards getting the NLEG test case working, and
as such it has changes to CompFlowProblemNLEnergyGrowth and other
CompFlowProblems. However, only the time-dependent Dirichlet BCs are
their assembly is tested and not the correctness of the NLEG solution
(and this not even with regression tests).

* CompFlow (e.g., NLEG) is not part of the regression test suite.

* There are no regression tests that would test Dirichlet BCs, neither
time-independent nor new time-dependent.

* While the previous way of constructing the data structures for BCs
worked with initial uniform refinement, the current time-dependent one
does not yet work with it.