Use unique triggers signaling to LHS and RHS
In this commit we make the SDAG trigger, bc_complete(), different for
consumed by LHS and for RHS. Previously, this was the same trigger, but
used in a multiple fashion and thus was triggered twice in
LinSysMerger::addbc(). Note that the meaning of the trigger
bc_complete() was that "the BC data structure LinSysMerger::m_bc is
ready for consumption" and since it was triggered twice, it meant that
it can be consumed by either the LHS or the RHS, setting the BCs.

Now we call the triggers for two use-cases differently,
bc_complete_rhs() and bc_complete_lhs(). This is because while
previously addbc() was called during setup as well as during time
stepping, now we only call it during time stepping. Since the LHS is not
regenerated during time stepping, putting in two bc_complete() triggers
now erroneously puts in one-too-many messages during time stepping for t
> 0. (It correctly puts in two messages during the first time step at t
= 0 at which we need to generate both LHS and RHS, but we only need a
single message during later time steps, for t > 0, since the LHS is not
regenerated anymore). Keeping the previous behavior (single trigger
called twice) while does not hang the run, it incorrectly never calls
the routine that is supposed to apply the BCs on the RHS, rhsbc() for t
> 0, because of (probably) the incorrect number of messages for t > 0
(for which there should only be a single message and not two).

By calling these two cases different the correct logic happens for both
t = 0 and t > 0. For t = 0, we need two messages, one for LHS and one
for LHS, and for t > 0, we need a message for RHS, this is now
bc_complete_rhs(), and while do not need a message for LHS, it does not
hurt to put in a differently-named one as bc_complete_lhs(), as it will
simply be never consumed for t > 0.

See also
https://github.com/quinoacomputing/quinoa/pull/146#discussion_r124055283
and
https://github.com/quinoacomputing/quinoa/commit/a156faa5cc60ba26354a28a96055026b39bd46b4#commitcomment-22773465
and the discussion at
https://lists.cs.illinois.edu/lists/arc/charm/2016-09/msg00002.html and
at https://lists.cs.illinois.edu/lists/arc/charm/2017-06/msg00013.html.

And no, I'm still not sure that I got this right... just documenting
what I find...