Add Taylor-Green regression tests
This is a slight modification of the Taylor-Green problem: it is made
stationary and it solves the Euler equations. For the original
Taylor-Green problem and its analytical solution see

G.I. Taylor, A.E. Green, "Mechanism of the Production of Small Eddies
from Large Ones", Proc. R. Soc. Lond. A 1937 158 499-521; DOI:
10.1098/rspa.1937.0036. Published 3 February 1937

For the modified one (the one that is implemented) see

Waltz, et. al, "Verification of a three-dimensional unstructured finite
element method using analytic and manufactured solutions", Computers and
Fluids, 2013, Vol.81, pp.57-67.