Correct asynchronous PDF collection in walker
This commit slightly changes the asynchronous logic and the estimation
of PDFs in walker. This is required for correctness using Charm++'s
randomized message queue, which exposed this problem.

Background and details:

The steps taken to numerically estimate statistical moments and
probability density functions (PDF) from particles, distributed across
CPUs and Charm++ chares, are as follows:

Integrator objects, instances of a Charm++ chare array, start by
accumulating sums for ordinary and central statistical moments as well
as ordinary and central PDFs. When partial estimation is done, they
forward their result to Collector objects which are Charm++ chare groups
and thus are guaranteed to be one per PE. Collector objects do partial
accumulation (of partially accumulate sums) on their PE, collecting from
Integrators that happen to be on their PE. Collectors then forward these
partial sums to their host, Distributor, a single Charm++ chare object,
which finishes the final step of the estimation. This happens for a
user-specified number and type of ordinary and central moments and for a
user-specified number and type of ordinary and central PDFs (and joint
PDFs) using custom typed Charm++ reductions. Central moments and central
PDFs can only be estimated when ordinary moments are available.

Previously, moments and PDFs were estimated separately, only enforcing
(via Charm++ SDAG) that ordinary moments must be available for before
the accumulation step for the central moments and central PDFs can be
started. This is correct and worked well with non-randomized queues, but
running with Charm++'s randomized queues, a problem was exposed in which
some messages as part of multiple reduction messages of multiple
reductions (probably) happened in unspecified order, which is not
allowed with the target of the reduction being the same single Charm++
chare.

This commit fixes that situation by enforcing that ordinary PDFs start
being accumulated immediately after ordinary moments, instead of
enforcing this logic via Charm++ SDAG. This basically means that the
functionality in Integrator::accumulateOrd() and
Integrator::accumlateOrdPDF() are combined and serialized. When both of
the ordinary moments and ordinary PDFs are partially accumulated, the
partial sums are sent to Collector via Collector::chareOrd(), in which
we now spawn the two reductions (one for moments, and another one for
PDFs) to Distributor after each other. This is done by also collapsing
the functionality in Collector::chareOrd() and Collector::chareOrdPDF().
Since Collector::chareOrd() spawns the reductions always in the order of
moments first and PDFs second, the required specific ordering is now
always maintained.