Correct asynchronous logic before partitioning in inciter
This commit slightly changes the asynchronous logic in inciter. This is
required for correctness using Charm++'s randomized message queue, which
exposed this problem.

Background and details:

A couple of the steps taken before mesh partitioning is (1) reading the
mesh graph from file, and (2) depending on whether a geometric or a
graph-based partitioner is selected by the user, computing the cell
centroids (required for geometric partitioners). When (1) is done,
Partitioners (Charm++group, one per PE) signal Transporter (single
Charm++ chare) via a reduction the fact that the graph has been read in
from file and sum up the total number of mesh cells across the whole
problem in Transporter::load(). Based on the total number of mesh cells
the computational load is computed in Transporter::load(). This sets the
number of worker chares across the problem, which can of course, be
larger than the number of PEs (if the problem is over-decomposed). The
number of chares is stored in Transporter::m_nchare and is set in
Transporter::load(). Independent of whether step (2) above happens or
not, a reduction to Transporter::partition() signals that the Carrier
worker chares are ready for mesh partitioning. Since partitioning
requires the number of chares, Carrier::partition() should only be
called when Transporter::m_nchare has been set. Previously this latter
logic was not in place, we just got lucky, but Charm++'s randomized
queue now exposed this problem.

The solution this commit puts in is to require that both reductions in
the above two steps complete before Partitioner::partition(), a
broadcast, is called. This is done via adding to Transporter's SDAG
logic: now both load_complete() and part_complete() bust be called
before partitioning can start. As a result Transporter::partition() has
now been renamed to Transporter::part(), which no longer calls
Partitioner::partition() but signals the runtime system by calling
part_complete(). The completion of the other required step is signaled
in Transporter::load() right after m_nchare is set by calling
load_complete().