Correct asynchronous logic during mesh nodal volume and statistics communication in inciter
This commit slightly changes the asynchronous logic in inciter. This is
required for correctness using Charm++'s randomized message queue, which
exposed this problem.

Background:

In Carrier (instances of inciter's Charm++ array worker chare), after
reading in the mesh coordinates for the mesh chunk held by a Carrier,
mesh cell volumes associated to nodes must be computed. This is required
for later operations when solving PDEs as well as reporting on mesh cell
volume statistics. There are three types of volumes computed:

(1) Nodal volumes, stored in Carrier:m_vol. This is the volume of the
mesh associated to nodes of owned elements (sum of surrounding cell
volumes / 4) with contributions from other chares on chare-boundaries.

(2) Nodal volumes, stored in Carrier::m_v. This is very similar to
m_vol, but this does not contain contributions from other chares on
chare boundaries. (On nodes that are not on chare boundaries, m_v =
m_vol.)

(3) Total volume, not stored in the state of Carrier but computed as a
diagnostic.

Obviously, all three volumes are used for different purposes.

After the mesh node coordinates have been read from file, in
Carrier::vol(), both first and second type of nodal volumes are
computed. While m_v does not need communication to be complete and
correct, those entries in m_vol that are associated to chare-boundary
nodes require communication across other chares that neighbor a given
chare. This is done via point-to-point communication via the entry
method Carrier::comvol(). When a Carrier chare has received all partial
volume contributions to its chare-boundary nodes in comvol (or if there
is nothing to communicate, i.e., in serial), a reduction is issued to
the host, Transporter::volcomplete().

The total volume is also computed in Carrier::vol() and need to be
communicated via a reduction to Transporter::vol() in which we globally
sum to arrive at the total mesh volume.

Another function, Carrier::stat(), computes various mesh cell
statistics, which also need to be communicated to Transporter, involving
four different types of reductions, computing minima, maxima, sums, and
histograms.

The problem:

All of the volumes and statistics, which need global reductions to
compute their final result, are independent and thus have previously
been spawned independently. However, Charm++ does not allow multiple
reductions that originate from the same array, target the same chare,
but spawned in an order that is not guaranteed to be always the same,
e.g., due to asynchronous logic happening differently, they are spawned
in different order.

The solution:

This commit makes sure the different global reductions computing the
first and the third type of volumes and the mesh statistics no longer
happen simultaneously.

This is done by effectively serializing the previously intermingled
multiple reductions by introducing artificial barriers (global
reductions and broadcasts) that would not be required for correctness. I
have not benchmarked the implications of these barriers. Note that
"serializing" here does not mean single-CPU, but that some of the
reductions are forced to happen after some other reductions have fully
completed.

Specific changes:

* m_vol is computed first, and the partial nodal volumes at
chare-boundary nodes are communicated to other Carrier objects. All
Carrier objects finish this by calling a reduction to
Transporter::volcomplete(), which starts the next step by issuing a
broadcast to Carrier::totalvol().

* The total volume is computed next in Carrier::totalvol() in which
Carrier objects sum up their partial mesh volume and contribute their
part to Transporter::totalvol(), which starts the next step by issuing a
broadcast to Carrier::stat().

* The final step is computing mesh cell statistics in Carrier::stat(),
in which after computing the mesh statistics four reductions to
different reduction entry methods of Transporter are spawned, which are
guaranteed to be issued in always the same order. Transporter::wait4stat
SDAG ensures that when all statistics are collected, the final mesh
statistics are output and the setup continues by calling Carrier::setup
in a broadcast fashion.

In summary, the above three steps all involve global reductions which
this commit forces to be strictly consecutive.